EC2 Name Generator
This code generates UID'd for assets like EC2 instances. The user inputs the department they belong to and the number of instances they want to name. After the conditions are met, names for the number of instances provided, are provided. If the user enters a department name that isn't approved, the program is aborted.